% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usgs.R
\docType{data}
\name{all_usgs_codes}
\alias{all_usgs_codes}
\title{All valid USGS codes}
\format{An object of class \code{integer} of length 205.}
\usage{
all_usgs_codes
}
\description{
All valid USGS codes
}
\keyword{datasets}
