% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{create_pattern_id_from_rgba_vec}
\alias{create_pattern_id_from_rgba_vec}
\title{Create the pattern id for a given hex colour for this package}
\usage{
create_pattern_id_from_rgba_vec(rgba_vec)
}
\arguments{
\item{rgba_vec}{4 element RGBA colour vector. Each element in range [0, 255]}
}
\value{
character string to use as a pattern ID
}
\description{
Create the pattern id for a given hex colour for this package
}
