% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-usgs.R
\name{create_usgs_pattern}
\alias{create_usgs_pattern}
\title{Create usgs pattern}
\usage{
create_usgs_pattern(
  usgs_code = 601,
  spacing = 20,
  fill = "#ffffff",
  alpha = 1,
  angle = 0,
  id = NULL
)
}
\arguments{
\item{usgs_code}{USGS code for a pattern}

\item{spacing}{size of pattern}

\item{fill}{fill colour}

\item{alpha}{default: 1.0 (opaque)}

\item{angle}{angle}

\item{id}{id to use for pattern. default: NULL means to generate automatically}
}
\value{
character string containing an svg <pattern>
}
\description{
Create usgs pattern
}
