% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{decode_pattern_from_rgba_vec}
\alias{decode_pattern_from_rgba_vec}
\title{Decode a 4-element RGBA colour vector as an SVG pattern}
\usage{
decode_pattern_from_rgba_vec(rgba_vec)
}
\arguments{
\item{rgba_vec}{4 element RGBA colour vector. Each element in range [0, 255]}
}
\value{
minisvg object
}
\description{
Decode a 4-element RGBA colour vector as an SVG pattern
}
