% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{encode_pattern_params_as_hex_colour}
\alias{encode_pattern_params_as_hex_colour}
\title{Encode a pattern as a hex colour}
\usage{
encode_pattern_params_as_hex_colour(...)
}
\arguments{
\item{...}{parameters to}
}
\value{
Return a hex colour representing the parameters
}
\description{
Encode a pattern as a hex colour
}
