% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{is_valid_pattern_encoding}
\alias{is_valid_pattern_encoding}
\title{Is the rgba_vec interpretable as a pattern?}
\usage{
is_valid_pattern_encoding(rgba_vec)
}
\arguments{
\item{rgba_vec}{4 element vector of RGBA values, each in the range [0, 255]}
}
\value{
logical value indicating whether or not thig RGBA colour is
        interpretable as a pattern in this package?
}
\description{
Is the rgba_vec interpretable as a pattern?
}
