% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usgs.R
\name{modify_usgs_pattern}
\alias{modify_usgs_pattern}
\title{Change the attributes of a pattern}
\usage{
modify_usgs_pattern(
  pattern,
  width,
  alpha = 1,
  fill = "#ffffff",
  angle = 0,
  height = width
)
}
\arguments{
\item{pattern}{svg text}

\item{width, height}{size in pixels}

\item{alpha}{opacity. default: 1. (opaque)}

\item{fill}{default: '#ffffff'}

\item{angle}{angle in degrees}
}
\value{
adjusted svg text
}
\description{
This function uses hand-crafted regular expressions to replace certain strings
in the given SVG with the values given as arguments.
}
\details{
This only really works because the SVG text that the USGS was derived from was
relatively regular in layout/naming.
}
