% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-usgs.R
\name{read_zipped_pattern_by_code}
\alias{read_zipped_pattern_by_code}
\title{Read an svg pattern from the zipped file given the USGS code}
\usage{
read_zipped_pattern_by_code(usgs_code = 601)
}
\arguments{
\item{usgs_code}{a USGS code for a pattern}
}
\value{
svg pattern object
}
\description{
This package comes with all SVG stored in a zip file.  This function takes
a USGS code and extracts that particular SVG from the zip file.
}
