% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-usgs.R
\name{create_usgs_pattern}
\alias{create_usgs_pattern}
\title{Create an SVG pattern from the USGS library}
\usage{
create_usgs_pattern(
  usgs_code = 601,
  spacing = 20,
  fill = "#ffffff",
  alpha = 1,
  angle = 0,
  id = NULL
)
}
\arguments{
\item{usgs_code}{USGS code for a pattern}

\item{spacing}{size of pattern}

\item{fill}{fill colour}

\item{alpha}{default: 1.0 (opaque)}

\item{angle}{angle}

\item{id}{id to use for pattern. default: NULL means to generate automatically}
}
\value{
minisvg::SVGPattern object
}
\description{
Create an SVG pattern from the USGS library
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the pattern and add to the SVG definitions
my_pattern <- create_usgs_pattern(usgs_code = 601, spacing = 100)
doc$defs(my_pattern)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  stroke = 'black',
  fill   = my_pattern
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
